GitProductivityDataWeavaite/requirements.txt
streamlit
openai
python-dotenv
langgraph
tavily-python
google-generativeai
langchain
PyPDF2
faiss-cpu
langchain_google_genai
langchain-community
langchain-core
langchain-openai
PyGithub
matplotlib
pandas
langchain
altair
weaviate-client
sentence-transformers==2.2.0
transformers
torch
huggingface-hub>=0.15.0



